Flesh out test harness.
